Extract vaccination related stuff from update.py
no other changes, just moved
